requirements.txt
django==4.2
djangorestframework
pika 
pandas
psycopg2-binary
python-decouple